Revert "删除内部的额外行为树"
This reverts commit df52713ab2dd60308637406b723ce79f9987928c.

发射完成

发射多个会销毁错

+++ Unity/Assets/_Scripts/Battle/SkillSystem/Other/SkillDesNodeData.cs
@@ -36,7 +36,7 @@ public class SkillDesNodeData : BuffNodeDataBase
     [HideLabel]
     [Title("技能CD", Bold = false)]
     [BsonDictionaryOptions(DictionaryRepresentation.ArrayOfArrays)]
-    public Dictionary<int, long> SkillCD = new();
+    public Dictionary<int, float> SkillCD = new();
 
     [TabGroup("基础信息")]
     [HideLabel]
